FUDGE-6.9.0 release
Changes include:
 - Updates for recent changes in matplotlib and numpy
 - Fixes for Windows installation
 - Improvements to physics checking code
 - Fixed a bug when generating transfer matrices for primary gammas
 - Fixed an issue in Merced (transfer matrix generating code) that potentially caused NaN results
 - Update energySpectrum.py to better support photo-nuclear evaluations

+++ bin/processProtare.py
@@ -138,6 +138,7 @@ parser.add_argument('--threads',type = int, default = 1,                        
 parser.add_argument('-v', '--verbose', action = 'count', default = 0,                                  help = 'Enable verbose output.')
 parser.add_argument('--reconAccuracy', type = float, default = reconAccuracyDefault,                   help = 'Accuracy for reconstructing resonances. Default is "%.1e".' % reconAccuracyDefault)
 parser.add_argument('--restart', action = 'store_true',                                                help = 'Continue previous incomplete processProtare run. If enabled, code checks for Merced output files from previous \nruns and if found reads them instead of rerunning Merced. Only impacts "-mg" option processing.')
+parser.add_argument('--skipMercedTar', action = 'store_true',                                          help = 'If present, the merced files are not tar-ed.')
 
 parser.add_argument('--preProcessLabel', type = str, default = None,                                   help = 'Label for style to process. If None, proc will pick the latest from the evaluated, and Realization styles.')
 
@@ -215,10 +216,10 @@ if args.temperatures is None:
         args.temperatures = [PQUModule.PQU(temperature, unit ).getValueAs(args.temperatureUnit) for temperature in temperatures]
 args.temperatures.sort()
 
-preProcessedStyles = reactionSuite.styles.preProcessingStyles()
-
 if args.cullProcessedData:
-    reactionSuite.cullProcessedData()
+    reactionSuite.cullProcessedData(include_reconstructed_data=True)
+
+preProcessedStyles = reactionSuite.styles.preProcessingStyles(include_reconstructed=True)
 
 for style in reactionSuite.styles:                 # Fail on detection of existing processed data.
     if not isinstance(style, preProcessedStyles): 
@@ -366,20 +367,42 @@ for temperatureIndex, temperatureValue in enumerate(args.temperatures):  # Heat 
         reactionSuite.processMultiGroup(heatedMultiGroupStyle, args.legendreMax, verbosity=args.verbose - 2, logFile=logFile, indent='    ', 
             additionalReactions=additionalReactions, workDir=str(workDir), restart=args.restart)
 
-        with tarfile.open(str(workDirTarFile), 'w:gz') as tar:
-            tar.add(workDir)
-        shutil.rmtree(str(workDir))
+        if not args.skipMercedTar:
+            with tarfile.open(str(workDirTarFile), 'w:gz') as tar:
+                tar.add(workDir)
+            shutil.rmtree(str(workDir))
 
         if args.UpScatter and reactionSuite.target != IDsPoPsModule.neutron:
             if args.verbose > 1:
                 print('  Processing multi group upscattering')
+
+            workDir = pathlib.Path('upscatter.work' ) / reactionSuite.projectile / reactionSuite.target / \
+                        ('T_%s_%s' % (temperatureValue, args.temperatureUnit.replace(' ', '').replace(os.sep, '_')))
+            workDirTarFile = pathlib.Path(str(workDir) + '.tar')
+            if workDirTarFile.exists():
+                if args.restart:
+                    tar = tarfile.open(str(workDirTarFile))
+                    tar.extractall()
+                else:
+                    workDirTarFile.unlink()
+            if workDir.exists() and not args.restart:
+                shutil.rmtree(str(workDir))
+            if not workDir.exists():
+                workDir.mkdir(parents=True)
+
             timerUp = timesModule.Times()
             heatUpscatterStyle = stylesModule.SnElasticUpScatter(args.prefixUpScatter + suffix, heatedMultiGroupStyle.label, date = dateTimeStr)
             reactionSuite.styles.add(heatUpscatterStyle)
             if temperatureValue > 0.0:
-                averageProductEnergy = reactionSuite.processSnElasticUpScatter(heatUpscatterStyle, args.legendreMax, verbosity=args.verbose - 2)  # FIXME need logfile arguments in this process call
+                averageProductEnergy = reactionSuite.processSnElasticUpScatter(
+                    heatUpscatterStyle, args.legendreMax, verbosity=args.verbose - 2,
+                    workDir = str(workDir), restart = args.restart)  # FIXME need logfile arguments in this process call
                 averageProductEnergyModelB.add(averageProductEnergy)
             logFile.write('  Upscatter correction\n    %s\n' % timerUp)
+
+            with tarfile.open(str(workDirTarFile), 'w:gz') as tar:
+                tar.add(workDir)
+            shutil.rmtree(str(workDir))
 
 if len(averageProductEnergyModelB) > 0:
     institution = institutionModule.Institution(deterministicTokensModule.pointwiseAverageProductEnergies)
